backend/requirements.txt
Django==4.1.3
django-colorfield==0.8.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
drf-yasg==1.21.4
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
